requirements.txt
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
idna==2.10
python-dotenv==0.15.0
requests==2.24.0
typer==0.3.2
urllib3==1.25.11
